refactor: fully migrate model list from ProviderSetting to ModelList
